@@ -7,35 +7,51 @@ import 'package:flutter_auto_form/src/models/field/field_context.dart';
 import 'package:flutter_auto_form/src/models/form.dart';
 import 'package:flutter_auto_form/src/widgets/theme.dart';
 
+abstract class AFFormStatefulWidget<T> extends StatefulWidget {
+  const AFFormStatefulWidget(
+      {super.key,
+      required this.formBuilder,
+      required this.onSubmitted,
+      required this.enableFinalAction,
+      this.enableSubmitFormWrapper,
+      this.handleErrorOnSubmit,
+      this.submitButton});
+
+  final T Function() formBuilder;
+
+  final Function(T form) onSubmitted;
+
+  final bool enableFinalAction;
+
+  final bool? enableSubmitFormWrapper;
+
+  final ValueChanged<String>? handleErrorOnSubmit;
+
+  final Widget Function(Function() submit)? submitButton;
+}
+
 /// The [AFFormState] allows to override and customize even more the behavior
 /// of the form widget's logic.
 ///
 /// Before considering extending this class, make sure that the [AFWidget] class
 /// does not satisfy your requirements!
-abstract class AFFormState<T extends StatefulWidget, G extends TemplateForm>
-    extends State<T> {
-  AFFormState({
-    required this.model,
-    this.enableFinalAction = true,
-    this.handleErrorOnSubmit,
-    this.enableSubmitFormWrapper,
-  });
-
+abstract class AFFormState<T extends AFFormStatefulWidget<G>,
+    G extends TemplateForm> extends State<T> {
   /// The [TemplateForm] that will be used as the blueprint of this class.
-  final G model;
+  late final G model;
 
   /// Whether to submit the form when the user clicks on the last field's [TextInputAction].
-  final bool enableFinalAction;
+  late final bool enableFinalAction;
 
   /// Whether to use wrap the submit function inside the [AFThemeData.submitFormWrapper]
   /// function. This can be useful to display a loading dialog while submitting
   /// the data.
-  final bool? enableSubmitFormWrapper;
+  late final bool? enableSubmitFormWrapper;
 
   /// A callback that will be triggered whenever the [submitForm] method is called
   /// while the form is invalid. It exposes the first error returned by one of the
   /// form's field ([Field]
-  final ValueChanged<String>? handleErrorOnSubmit;
+  late final ValueChanged<String>? handleErrorOnSubmit;
 
   /// Whether to display each field's respective error (if there is one) even if
   /// the user did not interact with any of these fields.
@@ -50,6 +66,14 @@ abstract class AFFormState<T extends StatefulWidget, G extends TemplateForm>
   @override
   void initState() {
     super.initState();
+
+    model = widget.formBuilder();
+
+    enableFinalAction = widget.enableFinalAction;
+
+    enableSubmitFormWrapper = widget.enableSubmitFormWrapper;
+
+    handleErrorOnSubmit = widget.handleErrorOnSubmit;
 
     FieldContext? previousFieldContext;
 
@@ -106,11 +130,11 @@ abstract class AFFormState<T extends StatefulWidget, G extends TemplateForm>
     });
 
     if (model.isComplete()) {
-      final bool _enabledSubmitFormWrapper = enableSubmitFormWrapper ??
+      final bool enabledSubmitFormWrapper = enableSubmitFormWrapper ??
           AFTheme.of(context).enableSubmitFormWrapper;
 
       try {
-        if (_enabledSubmitFormWrapper && enableFinalAction) {
+        if (enabledSubmitFormWrapper && enableFinalAction) {
           await theme.submitFormWrapper(
             context: context,
             future: submit(model),
